广告SDK接如中

+++ app/src/main/java/com/qubuxing/qbx/http/beans/JsonEvent.kt
@@ -1,5 +1,5 @@
 package com.qubuxing.qbx.http.beans
 
 class JsonEvent {
-    var step: Long =0
+    var step: Int =0
 }
